fix(kernels): rounding kernels take half inputs; trunc compiles; the bank measures them; probe modes
floor / ceil / remainder refused half-precision inputs — Triton's rounding
functions take fp32/fp64 only ("Expected dtype ['fp32', 'fp64'] but got
fp16") — so every --triton run of the RNNT speech model (parakeet) stopped
at its first floor. A half value is rounded in fp32 (exact: every half is
an fp32 value and the integer result is representable) and stored back in
its own dtype; fp32 inputs take the same path unchanged. trunc could never
compile (no tl.math.trunc) and is floor-or-ceil by sign. floor, ceil, trunc,
remainder join the reference bank (0 ULP on the rounding ops; remainder
measured at its quotient discontinuities). parakeet --triton: runs, no
torch at exit, transcript correct.

tools/r33_sys_modules_probe.py --warm: the serving warm path in-process
(the daemon's InferenceEngine, two requests) — TinyLlama under the Triton
engine: no torch at exit. tools/precision_zoo_campaign.py --probe [--src]:
the R33 lever over the whole zoo (one complete --triton request per model
under the probe, from a frozen worktree), with its own table.

+++ src/neurobrix/kernels/ops/remainder.py
@@ -27,7 +27,7 @@ def remainder_forward_kernel(
 
     x = tl.load(x_ptr + offset, mask=mask)
     y = tl.load(y_ptr + offset, mask=mask)
-    result = x - tl.math.floor(x / y) * y
+    result = x - tl.math.floor((x / y).to(tl.float32)).to(x.dtype) * y
     tl.store(output_ptr + offset, result, mask=mask)
 
 @triton.jit
@@ -42,5 +42,5 @@ def remainder_scalar_kernel(
     mask = offset < n_elements
 
     x = tl.load(x_ptr + offset, mask=mask)
-    result = x - tl.math.floor(x / divisor) * divisor
+    result = x - tl.math.floor((x / divisor).to(tl.float32)).to(x.dtype) * divisor
     tl.store(output_ptr + offset, result, mask=mask)
